+++ lib/ui/views/new_ib/components/new_ib_content.dart
@@ -1,15 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:mobile_app/ib_theme.dart';
 import 'package:mobile_app/locator.dart';
 import 'package:mobile_app/models/ib/ib_recommendation.dart';
 import 'package:mobile_app/models/ib/new_ib_drawer_data.dart';
-import 'package:mobile_app/models/ib/new_ib_home_data.dart';
-import 'package:mobile_app/models/ib/new_ib_chapter_index_data.dart';
-import 'package:mobile_app/models/ib/new_ib_topic_data.dart';
 import 'package:mobile_app/services/API/disqus_api.dart';
 import 'package:mobile_app/viewmodels/ib/new_ib_landing_viewmodel.dart';
-import 'package:mobile_app/ui/views/new_ib/components/binary_simulator_widget.dart';
-import 'package:mobile_app/ui/views/new_ib/components/quiz_widget.dart';
+import 'package:mobile_app/ui/views/new_ib/pages/home/home_page.dart';
+import 'package:mobile_app/ui/views/new_ib/pages/chapter/chapter_index_page.dart';
+import 'package:mobile_app/ui/views/new_ib/pages/topic/topic_page.dart';
 
 class NewIbContent extends StatefulWidget {
   final NewIbLandingViewModel model;
@@ -55,6 +52,15 @@ class _NewIbContentState extends State<NewIbContent> {
         widget.model.selectedChapter != null &&
         !widget.model.isHome) {
       print('NewIbContent: Calling _loadRecommendations from didUpdateWidget');
+      _loadRecommendations();
+    }
+    // Also load recommendations if we just switched to chapter index page
+    if (widget.model.topicData == null && 
+        oldWidget.model.topicData != null &&
+        widget.model.selectedChapter != null &&
+        !widget.model.isHome &&
+        _recommendations.isEmpty) {
+      print('NewIbContent: Calling _loadRecommendations after returning from topic');
       _loadRecommendations();
     }
   }
@@ -125,1815 +131,48 @@ class _NewIbContentState extends State<NewIbContent> {
   Widget build(BuildContext context) {
     // Show home page if no chapter selected or home is selected
     if (widget.model.isHome || widget.model.selectedChapter == null) {
-      return _buildHomePage(context);
+      return IbHomePage(
+        model: widget.model,
+        onNavigate: widget.onNavigate,
+      );
     }
 
     // Show topic page if topic data is available
     if (widget.model.topicData != null) {
-      return _buildTopicPage(context, widget.model.topicData!);
-    }
-
-    // Show chapter page
-    return _buildChapterPage(context, widget.model.selectedChapter!);
-  }
-
-  Widget _buildHomePage(BuildContext context) {
-    // Show loading state
-    if (widget.model.isBusy(widget.model.NEW_IB_FETCH_HOME)) {
-      return const Center(
-        child: CircularProgressIndicator(),
-      );
-    }
-
-    // Show error state
-    if (widget.model.isError(widget.model.NEW_IB_FETCH_HOME)) {
-      return Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              Icons.error_outline_rounded,
-              size: 64,
-              color: IbTheme.textColor(context).withAlpha(128),
-            ),
-            const SizedBox(height: 16),
-            Text(
-              'Failed to load home page',
-              style: TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.w600,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-            const SizedBox(height: 8),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 32),
-              child: Text(
-                widget.model.errorMessageFor(widget.model.NEW_IB_FETCH_HOME) ?? 'Unknown error',
-                style: TextStyle(
-                  fontSize: 14,
-                  color: IbTheme.textColor(context).withAlpha(179),
-                ),
-                textAlign: TextAlign.center,
-              ),
-            ),
-            const SizedBox(height: 24),
-            ElevatedButton.icon(
-              onPressed: () => widget.model.fetchHomeData(),
-              icon: const Icon(Icons.refresh_rounded),
-              label: const Text('Retry'),
-              style: ElevatedButton.styleFrom(
-                backgroundColor: IbTheme.getPrimaryColor(context),
-                foregroundColor: Colors.white,
-              ),
-            ),
-          ],
-        ),
-      );
-    }
-
-    // Get home data
-    final homeData = widget.model.homeData;
-    if (homeData == null) {
-      return const Center(child: Text('No home data available'));
-    }
-
-    return SingleChildScrollView(
-      padding: const EdgeInsets.all(24),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          // Hero Section with Icon
-          Container(
-            padding: const EdgeInsets.all(32),
-            decoration: BoxDecoration(
-              color: IbTheme.getPrimaryColor(context).withAlpha(26),
-              borderRadius: BorderRadius.circular(20),
-            ),
-            child: Column(
-              children: [
-                Container(
-                  padding: const EdgeInsets.all(24),
-                  decoration: BoxDecoration(
-                    color: IbTheme.getPrimaryColor(context).withAlpha(51),
-                    shape: BoxShape.circle,
-                  ),
-                  child: Icon(
-                    Icons.menu_book_rounded,
-                    size: 56,
-                    color: IbTheme.getPrimaryColor(context),
-                  ),
-                ),
-                const SizedBox(height: 24),
-                Text(
-                  homeData.title,
-                  style: TextStyle(
-                    fontSize: 28,
-                    fontWeight: FontWeight.bold,
-                    color: IbTheme.primaryHeadingColor(context),
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-                const SizedBox(height: 12),
-                Text(
-                  homeData.subtitle,
-                  style: TextStyle(
-                    fontSize: 16,
-                    color: IbTheme.textColor(context).withAlpha(204),
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-              ],
-            ),
-          ),
-          
-          const SizedBox(height: 24),
-
-          // Content Box with all sections
-          Container(
-            padding: const EdgeInsets.all(20),
-            decoration: BoxDecoration(
-              color: IbTheme.textColor(context).withAlpha(13),
-              borderRadius: BorderRadius.circular(16),
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: homeData.content.sections.map((section) {
-                return _buildHomeSection(context, section);
-              }).toList(),
-            ),
-          ),
-
-          const SizedBox(height: 24),
-
-          // Feature Cards Grid (2x2)
-          Row(
-            children: [
-              Expanded(
-                child: _buildFeatureCard(
-                  context,
-                  Icons.school_rounded,
-                  'Learn',
-                  'Comprehensive digital logic tutorials',
-                ),
-              ),
-              const SizedBox(width: 16),
-              Expanded(
-                child: _buildFeatureCard(
-                  context,
-                  Icons.science_rounded,
-                  'Experiment',
-                  'Interactive circuit simulations',
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 16),
-          Row(
-            children: [
-              Expanded(
-                child: _buildFeatureCard(
-                  context,
-                  Icons.quiz_rounded,
-                  'Practice',
-                  'Quizzes and exercises',
-                ),
-              ),
-              const SizedBox(width: 16),
-              Expanded(
-                child: _buildFeatureCard(
-                  context,
-                  Icons.emoji_events_rounded,
-                  'Master',
-                  'Track your progress',
-                ),
-              ),
-            ],
-          ),
-
-          const SizedBox(height: 24),
-
-          // Getting Started Section
-          Container(
-            padding: const EdgeInsets.all(24),
-            decoration: BoxDecoration(
-              gradient: LinearGradient(
-                begin: Alignment.topLeft,
-                end: Alignment.bottomRight,
-                colors: [
-                  IbTheme.getPrimaryColor(context),
-                  IbTheme.getPrimaryColor(context).withAlpha(204),
-                ],
-              ),
-              borderRadius: BorderRadius.circular(20),
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(
-                  children: [
-                    Container(
-                      padding: const EdgeInsets.all(12),
-                      decoration: BoxDecoration(
-                        color: Colors.white.withAlpha(51),
-                        borderRadius: BorderRadius.circular(12),
-                      ),
-                      child: const Icon(
-                        Icons.rocket_launch_rounded,
-                        color: Colors.white,
-                        size: 28,
-                      ),
-                    ),
-                    const SizedBox(width: 16),
-                    const Text(
-                      'Getting Started',
-                      style: TextStyle(
-                        fontSize: 24,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.white,
-                      ),
-                    ),
-                  ],
-                ),
-                const SizedBox(height: 24),
-                _buildGettingStartedStep(
-                  context,
-                  '1',
-                  'Choose a Chapter',
-                  'Browse chapters from the menu',
-                ),
-                const SizedBox(height: 16),
-                _buildGettingStartedStep(
-                  context,
-                  '2',
-                  'Learn Concepts',
-                  'Read through interactive lessons',
-                ),
-                const SizedBox(height: 16),
-                _buildGettingStartedStep(
-                  context,
-                  '3',
-                  'Practice',
-                  'Test your knowledge with quizzes',
-                ),
-              ],
-            ),
-          ),
-
-          const SizedBox(height: 32),
-
-          // Footer
-          Center(
-            child: Text(
-              homeData.metadata.copyright,
-              style: TextStyle(
-                fontSize: 12,
-                color: IbTheme.textColor(context).withAlpha(128),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildHomeSection(BuildContext context, NewIbHomeSection section) {
-    switch (section.type) {
-      case 'heading':
-        if (section.level == 1) {
-          // Skip H1 as it's already in the hero section
-          return const SizedBox.shrink();
-        } else if (section.level == 2) {
-          return Padding(
-            padding: const EdgeInsets.only(top: 16, bottom: 12),
-            child: Text(
-              section.text ?? '',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-          );
-        } else if (section.level == 3) {
-          return Padding(
-            padding: const EdgeInsets.only(top: 12, bottom: 8),
-            child: Text(
-              section.text ?? '',
-              style: TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.w600,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-          );
-        }
-        return const SizedBox.shrink();
-
-      case 'subtitle':
-        // Skip subtitle as it's in the hero section
-        return const SizedBox.shrink();
-
-      case 'paragraph':
-        return Padding(
-          padding: const EdgeInsets.only(bottom: 16),
-          child: Text(
-            section.text ?? '',
-            style: TextStyle(
-              fontSize: 15,
-              height: 1.6,
-              color: IbTheme.textColor(context).withAlpha(204),
-            ),
-          ),
-        );
-
-      case 'divider':
-        return const Padding(
-          padding: EdgeInsets.symmetric(vertical: 12),
-          child: Divider(),
-        );
-
-      default:
-        return const SizedBox.shrink();
-    }
-  }
-
-  Widget _buildFeatureCard(
-    BuildContext context,
-    IconData icon,
-    String title,
-    String description,
-  ) {
-    return Container(
-      padding: const EdgeInsets.all(20),
-      decoration: BoxDecoration(
-        color: IbTheme.textColor(context).withAlpha(13),
-        borderRadius: BorderRadius.circular(16),
-        border: Border.all(
-          color: IbTheme.textColor(context).withAlpha(26),
-        ),
-      ),
-      child: Column(
-        children: [
-          Container(
-            padding: const EdgeInsets.all(16),
-            decoration: BoxDecoration(
-              color: IbTheme.getPrimaryColor(context).withAlpha(26),
-              borderRadius: BorderRadius.circular(12),
-            ),
-            child: Icon(
-              icon,
-              size: 32,
-              color: IbTheme.getPrimaryColor(context),
-            ),
-          ),
-          const SizedBox(height: 12),
-          Text(
-            title,
-            style: TextStyle(
-              fontSize: 16,
-              fontWeight: FontWeight.w600,
-              color: IbTheme.primaryHeadingColor(context),
-            ),
-            textAlign: TextAlign.center,
-          ),
-          const SizedBox(height: 6),
-          Text(
-            description,
-            style: TextStyle(
-              fontSize: 12,
-              color: IbTheme.textColor(context).withAlpha(179),
-            ),
-            textAlign: TextAlign.center,
-            maxLines: 2,
-            overflow: TextOverflow.ellipsis,
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildGettingStartedStep(
-    BuildContext context,
-    String number,
-    String title,
-    String description,
-  ) {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Container(
-          width: 36,
-          height: 36,
-          decoration: BoxDecoration(
-            color: Colors.white.withAlpha(51),
-            borderRadius: BorderRadius.circular(10),
-          ),
-          child: Center(
-            child: Text(
-              number,
-              style: const TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.bold,
-                color: Colors.white,
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(width: 16),
-        Expanded(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                title,
-                style: const TextStyle(
-                  fontSize: 16,
-                  fontWeight: FontWeight.w600,
-                  color: Colors.white,
-                ),
-              ),
-              const SizedBox(height: 4),
-              Text(
-                description,
-                style: TextStyle(
-                  fontSize: 14,
-                  color: Colors.white.withAlpha(204),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildChapterPage(BuildContext context, NewIbChapter chapter) {
-    // Load recommendations if not already loaded or loading
-    if (!_loadingRecommendations && _recommendations.isEmpty) {
-      // Use addPostFrameCallback to avoid calling setState during build
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        if (mounted && !_loadingRecommendations && _recommendations.isEmpty) {
-          print('NewIbContent: Calling _loadRecommendations from _buildChapterPage');
-          _loadRecommendations();
-        }
-      });
-    }
-
-    // Show loading state
-    if (widget.model.isBusy(widget.model.NEW_IB_FETCH_CHAPTER_INDEX)) {
-      return const Center(
-        child: CircularProgressIndicator(),
-      );
-    }
-
-    // Show error state - display "Coming Soon" instead
-    if (widget.model.isError(widget.model.NEW_IB_FETCH_CHAPTER_INDEX)) {
-      return Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              Icons.construction_rounded,
-              size: 80,
-              color: IbTheme.getPrimaryColor(context).withAlpha(128),
-            ),
-            const SizedBox(height: 24),
-            Text(
-              'Coming Soon',
-              style: TextStyle(
-                fontSize: 32,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-            const SizedBox(height: 12),
-            Text(
-              'This page is under development',
-              style: TextStyle(
-                fontSize: 16,
-                color: IbTheme.textColor(context).withAlpha(179),
-              ),
-            ),
-          ],
-        ),
-      );
-    }
-
-    // Get chapter index data
-    final chapterData = widget.model.chapterIndexData;
-    if (chapterData == null) {
-      return const Center(child: Text('No chapter data available'));
-    }
-
-    return SingleChildScrollView(
-      padding: const EdgeInsets.all(24),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          // Chapter Title and Description Box
-          Container(
-            padding: const EdgeInsets.all(24),
-            decoration: BoxDecoration(
-              color: IbTheme.getPrimaryColor(context).withAlpha(26),
-              borderRadius: BorderRadius.circular(16),
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  chapterData.title,
-                  style: TextStyle(
-                    fontSize: 28,
-                    fontWeight: FontWeight.bold,
-                    color: IbTheme.primaryHeadingColor(context),
-                  ),
-                ),
-                const SizedBox(height: 12),
-                Text(
-                  chapterData.description,
-                  style: TextStyle(
-                    fontSize: 16,
-                    height: 1.5,
-                    color: IbTheme.textColor(context).withAlpha(204),
-                  ),
-                ),
-              ],
-            ),
-          ),
-
-          const SizedBox(height: 24),
-
-          // Content sections
-          if (chapterData.content.sections.isNotEmpty)
-            Container(
-              padding: const EdgeInsets.all(20),
-              decoration: BoxDecoration(
-                color: IbTheme.textColor(context).withAlpha(13),
-                borderRadius: BorderRadius.circular(16),
-              ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: chapterData.content.sections.map((section) {
-                  return _buildChapterSection(context, section);
-                }).toList(),
-              ),
-            ),
-
-          const SizedBox(height: 24),
-
-          // Children/Sub-chapters list
-          if (chapterData.children.isNotEmpty) ...[
-            Text(
-              'Topics in this chapter',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-            const SizedBox(height: 16),
-            ...chapterData.children.map((child) => _buildChildTile(context, child)),
-          ],
-
-          const SizedBox(height: 32),
-
-          // Also on Interactive Book (Recommendations)
-          _buildRecommendationsSection(context),
-
-          const SizedBox(height: 32),
-
-          // Comments Section
-          _buildCommentsSection(context),
-
-          const SizedBox(height: 80),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildChapterSection(BuildContext context, NewIbChapterSection section) {
-    switch (section.type) {
-      case 'heading':
-        if (section.level == 1) {
-          return Padding(
-            padding: const EdgeInsets.only(bottom: 16),
-            child: Text(
-              section.text ?? '',
-              style: TextStyle(
-                fontSize: 24,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-          );
-        } else if (section.level == 2) {
-          return Padding(
-            padding: const EdgeInsets.only(top: 16, bottom: 12),
-            child: Text(
-              section.text ?? '',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-          );
-        } else if (section.level == 3) {
-          return Padding(
-            padding: const EdgeInsets.only(top: 12, bottom: 8),
-            child: Text(
-              section.text ?? '',
-              style: TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.w600,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-          );
-        }
-        return const SizedBox.shrink();
-
-      case 'paragraph':
-        return Padding(
-          padding: const EdgeInsets.only(bottom: 16),
-          child: Text(
-            section.text ?? '',
-            style: TextStyle(
-              fontSize: 15,
-              height: 1.6,
-              color: IbTheme.textColor(context).withAlpha(204),
-            ),
-          ),
-        );
-
-      default:
-        return const SizedBox.shrink();
-    }
-  }
-
-  Widget _buildChildTile(BuildContext context, NewIbChapterChild child) {
-    return Container(
-      margin: const EdgeInsets.only(bottom: 12),
-      decoration: BoxDecoration(
-        color: IbTheme.textColor(context).withAlpha(13),
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(
-          color: IbTheme.textColor(context).withAlpha(26),
-        ),
-      ),
-      child: InkWell(
-        onTap: () {
-          // Navigate to topic page by fetching topic data
-          widget.model.fetchTopicData(child.path);
-        },
-        borderRadius: BorderRadius.circular(12),
-        child: Padding(
-          padding: const EdgeInsets.all(16),
-          child: Row(
-            children: [
-              Container(
-                padding: const EdgeInsets.all(12),
-                decoration: BoxDecoration(
-                  color: IbTheme.getPrimaryColor(context).withAlpha(26),
-                  borderRadius: BorderRadius.circular(10),
-                ),
-                child: Icon(
-                  Icons.article_rounded,
-                  size: 24,
-                  color: IbTheme.getPrimaryColor(context),
-                ),
-              ),
-              const SizedBox(width: 16),
-              Expanded(
-                child: Text(
-                  child.title,
-                  style: TextStyle(
-                    fontSize: 16,
-                    fontWeight: FontWeight.w500,
-                    color: IbTheme.primaryHeadingColor(context),
-                  ),
-                ),
-              ),
-              Icon(
-                Icons.arrow_forward_ios_rounded,
-                size: 16,
-                color: IbTheme.textColor(context).withAlpha(128),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildRecommendationsSection(BuildContext context) {
-    print('NewIbContent: _buildRecommendationsSection called');
-    print('NewIbContent: _loadingRecommendations: $_loadingRecommendations');
-    print('NewIbContent: _recommendations.length: ${_recommendations.length}');
-    
-    if (_loadingRecommendations) {
-      return Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              Icon(
-                Icons.auto_stories_rounded,
-                color: IbTheme.getPrimaryColor(context),
-                size: 28,
-              ),
-              const SizedBox(width: 12),
-              Text(
-                'Also on Interactive Book',
-                style: TextStyle(
-                  fontSize: 24,
-                  fontWeight: FontWeight.bold,
-                  color: IbTheme.primaryHeadingColor(context),
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 16),
-          const Center(child: CircularProgressIndicator()),
-        ],
-      );
-    }
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            Icon(
-              Icons.auto_stories_rounded,
-              color: IbTheme.getPrimaryColor(context),
-              size: 28,
-            ),
-            const SizedBox(width: 12),
-            Text(
-              'Also on Interactive Book',
-              style: TextStyle(
-                fontSize: 24,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 16),
-        if (_recommendations.isEmpty)
-          Center(
-            child: Padding(
-              padding: const EdgeInsets.all(24.0),
-              child: Text(
-                'No recommendations available',
-                style: TextStyle(
-                  fontSize: 14,
-                  color: IbTheme.textColor(context).withAlpha(128),
-                ),
-              ),
-            ),
-          )
-        else
-          SizedBox(
-            height: 140,
-            child: ListView.builder(
-              scrollDirection: Axis.horizontal,
-              itemCount: _recommendations.length > 5 ? 5 : _recommendations.length,
-              itemBuilder: (context, index) {
-                final rec = _recommendations[index];
-                return _buildRecommendationCard(context, rec);
-              },
-            ),
-          ),
-      ],
-    );
-  }
-
-  Widget _buildRecommendationCard(
-    BuildContext context,
-    IbRecommendation recommendation,
-  ) {
-    // Format date
-    String formattedDate = '';
-    if (recommendation.createdAt != null) {
-      try {
-        final date = DateTime.parse(recommendation.createdAt!);
-        formattedDate =
-            '${date.year}-${date.month.toString().padLeft(2, '0')}-${date.day.toString().padLeft(2, '0')}';
-      } catch (e) {
-        formattedDate = '';
-      }
-    }
-    return Container(
-      width: 280,
-      height: 140,
-      margin: const EdgeInsets.only(right: 12),
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(color: IbTheme.textColor(context).withAlpha(26)),
-        color: IbTheme.getPrimaryColor(context).withAlpha(51),
-      ),
-      child: ClipRRect(
-        borderRadius: BorderRadius.circular(12),
-        child: Stack(
-          fit: StackFit.expand,
-          children: [
-            // Background image
-            if (recommendation.image != null)
-              Image.network(
-                recommendation.image!,
-                fit: BoxFit.cover,
-                errorBuilder: (context, error, stackTrace) {
-                  return Container(
-                    color: IbTheme.getPrimaryColor(context).withAlpha(51),
-                  );
-                },
-                loadingBuilder: (context, child, loadingProgress) {
-                  if (loadingProgress == null) return child;
-                  return Container(
-                    color: IbTheme.getPrimaryColor(context).withAlpha(51),
-                    child: Center(
-                      child: CircularProgressIndicator(
-                        value: loadingProgress.expectedTotalBytes != null
-                            ? loadingProgress.cumulativeBytesLoaded /
-                                loadingProgress.expectedTotalBytes!
-                            : null,
-                      ),
-                    ),
-                  );
-                },
-              ),
-            // Dark overlay
-            Container(
-              decoration: BoxDecoration(
-                gradient: LinearGradient(
-                  begin: Alignment.topCenter,
-                  end: Alignment.bottomCenter,
-                  colors: [
-                    Colors.black.withOpacity(0.3),
-                    Colors.black.withOpacity(0.8),
-                  ],
-                ),
-              ),
-            ),
-            // Content
-            InkWell(
-              onTap: () {},
-              child: Padding(
-                padding: const EdgeInsets.all(16),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      recommendation.title,
-                      style: const TextStyle(
-                        fontSize: 16,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.white,
-                      ),
-                      maxLines: 2,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                    const SizedBox(height: 8),
-                    Row(
-                      children: [
-                        if (formattedDate.isNotEmpty) ...[
-                          const Icon(
-                            Icons.calendar_today,
-                            size: 12,
-                            color: Colors.white70,
-                          ),
-                          const SizedBox(width: 4),
-                          Text(
-                            formattedDate,
-                            style: const TextStyle(
-                              fontSize: 11,
-                              color: Colors.white70,
-                            ),
-                          ),
-                        ],
-                        if (formattedDate.isNotEmpty && recommendation.posts != null)
-                          const SizedBox(width: 12),
-                        if (recommendation.posts != null) ...[
-                          const Icon(
-                            Icons.comment,
-                            size: 12,
-                            color: Colors.white70,
-                          ),
-                          const SizedBox(width: 4),
-                          Text(
-                            '${recommendation.posts} comment${recommendation.posts! > 1 ? 's' : ''}',
-                            style: const TextStyle(
-                              fontSize: 11,
-                              color: Colors.white70,
-                            ),
-                          ),
-                        ],
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildCommentsSection(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            Icon(
-              Icons.comment_rounded,
-              color: IbTheme.getPrimaryColor(context),
-              size: 28,
-            ),
-            const SizedBox(width: 12),
-            Text(
-              'Comments',
-              style: TextStyle(
-                fontSize: 24,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 16),
-        _buildCommentInput(context),
-        const SizedBox(height: 24),
-        _buildCommentsList(context),
-      ],
-    );
-  }
-
-  Widget _buildTopicPage(BuildContext context, NewIbTopicData topicData) {
-    // Load recommendations if not already loaded or loading
-    if (!_loadingRecommendations && _recommendations.isEmpty) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        if (mounted && !_loadingRecommendations && _recommendations.isEmpty) {
-          print('NewIbContent: Calling _loadRecommendations from _buildTopicPage');
-          _loadRecommendations();
-        }
-      });
-    }
-
-    // Show loading state
-    if (widget.model.isBusy(widget.model.NEW_IB_FETCH_TOPIC)) {
-      return const Center(
-        child: CircularProgressIndicator(),
-      );
-    }
-
-    // Show error state - display "Coming Soon" instead
-    if (widget.model.isError(widget.model.NEW_IB_FETCH_TOPIC)) {
-      return Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              Icons.construction_rounded,
-              size: 80,
-              color: IbTheme.getPrimaryColor(context).withAlpha(128),
-            ),
-            const SizedBox(height: 24),
-            Text(
-              'Coming Soon',
-              style: TextStyle(
-                fontSize: 32,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-            const SizedBox(height: 12),
-            Text(
-              'This page is under development',
-              style: TextStyle(
-                fontSize: 16,
-                color: IbTheme.textColor(context).withAlpha(179),
-              ),
-            ),
-          ],
-        ),
-      );
-    }
-
-    return SingleChildScrollView(
-      padding: const EdgeInsets.all(24),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          // Topic Title Box
-          Container(
-            padding: const EdgeInsets.all(24),
-            decoration: BoxDecoration(
-              color: IbTheme.getPrimaryColor(context).withAlpha(26),
-              borderRadius: BorderRadius.circular(16),
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  topicData.title,
-                  style: TextStyle(
-                    fontSize: 28,
-                    fontWeight: FontWeight.bold,
-                    color: IbTheme.primaryHeadingColor(context),
-                  ),
-                ),
-                if (topicData.metadata.author != null) ...[
-                  const SizedBox(height: 8),
-                  Row(
-                    children: [
-                      Icon(
-                        Icons.person_outline_rounded,
-                        size: 16,
-                        color: IbTheme.textColor(context).withAlpha(179),
-                      ),
-                      const SizedBox(width: 6),
-                      Text(
-                        topicData.metadata.author!,
-                        style: TextStyle(
-                          fontSize: 14,
-                          color: IbTheme.textColor(context).withAlpha(179),
-                        ),
-                      ),
-                    ],
-                  ),
-                ],
-              ],
-            ),
-          ),
-
-          const SizedBox(height: 24),
-
-          // Content sections
-          Container(
-            padding: const EdgeInsets.all(20),
-            decoration: BoxDecoration(
-              color: IbTheme.textColor(context).withAlpha(13),
-              borderRadius: BorderRadius.circular(16),
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: topicData.content.sections.map((section) {
-                return _buildTopicSection(context, section);
-              }).toList(),
-            ),
-          ),
-
-          const SizedBox(height: 24),
-
-          // Key Concepts
-          if (topicData.keyConcepts.isNotEmpty) ...[
-            Container(
-              padding: const EdgeInsets.all(20),
-              decoration: BoxDecoration(
-                color: IbTheme.getPrimaryColor(context).withAlpha(26),
-                borderRadius: BorderRadius.circular(16),
-              ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Row(
-                    children: [
-                      Icon(
-                        Icons.lightbulb_outline_rounded,
-                        color: IbTheme.getPrimaryColor(context),
-                        size: 24,
-                      ),
-                      const SizedBox(width: 12),
-                      Text(
-                        'Key Concepts',
-                        style: TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.bold,
-                          color: IbTheme.primaryHeadingColor(context),
-                        ),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 16),
-                  ...topicData.keyConcepts.map((concept) {
-                    return Padding(
-                      padding: const EdgeInsets.only(bottom: 12),
-                      child: Row(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Container(
-                            margin: const EdgeInsets.only(top: 6),
-                            width: 6,
-                            height: 6,
-                            decoration: BoxDecoration(
-                              color: IbTheme.getPrimaryColor(context),
-                              shape: BoxShape.circle,
-                            ),
-                          ),
-                          const SizedBox(width: 12),
-                          Expanded(
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: [
-                                Text(
-                                  concept.concept,
-                                  style: TextStyle(
-                                    fontSize: 16,
-                                    fontWeight: FontWeight.w600,
-                                    color: IbTheme.primaryHeadingColor(context),
-                                  ),
-                                ),
-                                const SizedBox(height: 4),
-                                Text(
-                                  concept.description,
-                                  style: TextStyle(
-                                    fontSize: 14,
-                                    color: IbTheme.textColor(context).withAlpha(204),
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                        ],
-                      ),
-                    );
-                  }).toList(),
-                ],
-              ),
-            ),
-            const SizedBox(height: 24),
-          ],
-
-          // Related Topics
-          if (topicData.relatedTopics.isNotEmpty) ...[
-            Text(
-              'Related Topics',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-            const SizedBox(height: 16),
-            ...topicData.relatedTopics.map((relatedTopic) {
-              return Container(
-                margin: const EdgeInsets.only(bottom: 12),
-                decoration: BoxDecoration(
-                  color: IbTheme.textColor(context).withAlpha(13),
-                  borderRadius: BorderRadius.circular(12),
-                  border: Border.all(
-                    color: IbTheme.textColor(context).withAlpha(26),
-                  ),
-                ),
-                child: InkWell(
-                  onTap: () {
-                    widget.model.fetchTopicData(relatedTopic.path);
-                  },
-                  borderRadius: BorderRadius.circular(12),
-                  child: Padding(
-                    padding: const EdgeInsets.all(16),
-                    child: Row(
-                      children: [
-                        Container(
-                          padding: const EdgeInsets.all(12),
-                          decoration: BoxDecoration(
-                            color: IbTheme.getPrimaryColor(context).withAlpha(26),
-                            borderRadius: BorderRadius.circular(10),
-                          ),
-                          child: Icon(
-                            Icons.article_rounded,
-                            size: 24,
-                            color: IbTheme.getPrimaryColor(context),
-                          ),
-                        ),
-                        const SizedBox(width: 16),
-                        Expanded(
-                          child: Text(
-                            relatedTopic.title,
-                            style: TextStyle(
-                              fontSize: 16,
-                              fontWeight: FontWeight.w500,
-                              color: IbTheme.primaryHeadingColor(context),
-                            ),
-                          ),
-                        ),
-                        Icon(
-                          Icons.arrow_forward_ios_rounded,
-                          size: 16,
-                          color: IbTheme.textColor(context).withAlpha(128),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              );
-            }).toList(),
-            const SizedBox(height: 32),
-          ],
-
-          // Recommendations
-          _buildRecommendationsSection(context),
-
-          const SizedBox(height: 32),
-
-          // Comments
-          _buildCommentsSection(context),
-
-          const SizedBox(height: 80),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildTopicSection(BuildContext context, dynamic section) {
-    final sectionType = section['type'] as String?;
-
-    switch (sectionType) {
-      case 'heading':
-        final level = section['level'] as int? ?? 1;
-        final text = section['text'] as String? ?? '';
-        
-        if (level == 1) {
-          return Padding(
-            padding: const EdgeInsets.only(bottom: 16),
-            child: Text(
-              text,
-              style: TextStyle(
-                fontSize: 24,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-          );
-        } else if (level == 2) {
-          return Padding(
-            padding: const EdgeInsets.only(top: 16, bottom: 12),
-            child: Text(
-              text,
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.bold,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-          );
-        } else if (level == 3) {
-          return Padding(
-            padding: const EdgeInsets.only(top: 12, bottom: 8),
-            child: Text(
-              text,
-              style: TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.w600,
-                color: IbTheme.primaryHeadingColor(context),
-              ),
-            ),
-          );
-        }
-        return const SizedBox.shrink();
-
-      case 'section':
-        final heading = section['heading'];
-        final headingMap = heading != null 
-            ? (heading is Map<String, dynamic> 
-                ? heading 
-                : Map<String, dynamic>.from(heading as Map))
-            : null;
-        final content = section['content'] as List<dynamic>? ?? [];
-        
-        return Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            if (headingMap != null) _buildTopicSection(context, headingMap),
-            ...content.map((item) => _buildTopicContentItem(context, item)).toList(),
-          ],
-        );
-
-      default:
-        return const SizedBox.shrink();
-    }
-  }
-
-  Widget _buildTopicContentItem(BuildContext context, dynamic item) {
-    final itemType = item['type'] as String?;
-
-    switch (itemType) {
-      case 'paragraph':
-        final text = item['text'] as String? ?? '';
-        return Padding(
-          padding: const EdgeInsets.only(bottom: 16),
-          child: Text(
-            text,
-            style: TextStyle(
-              fontSize: 15,
-              height: 1.6,
-              color: IbTheme.textColor(context).withAlpha(204),
-            ),
-          ),
-        );
-
-      case 'example':
-        final title = item['title'] as String? ?? 'Example';
-        final dataRaw = item['data'];
-        final data = dataRaw != null
-            ? (dataRaw is Map<String, dynamic>
-                ? dataRaw
-                : Map<String, dynamic>.from(dataRaw as Map))
-            : <String, dynamic>{};
-        return Container(
-          margin: const EdgeInsets.only(bottom: 16),
-          padding: const EdgeInsets.all(16),
-          decoration: BoxDecoration(
-            color: IbTheme.getPrimaryColor(context).withAlpha(26),
-            borderRadius: BorderRadius.circular(12),
-            border: Border.all(
-              color: IbTheme.getPrimaryColor(context).withAlpha(51),
-            ),
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                children: [
-                  Icon(
-                    Icons.lightbulb_outline_rounded,
-                    size: 20,
-                    color: IbTheme.getPrimaryColor(context),
-                  ),
-                  const SizedBox(width: 8),
-                  Expanded(
-                    child: Text(
-                      title,
-                      style: TextStyle(
-                        fontSize: 16,
-                        fontWeight: FontWeight.w600,
-                        color: IbTheme.primaryHeadingColor(context),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 12),
-              ...data.entries.map((entry) {
-                return Padding(
-                  padding: const EdgeInsets.only(bottom: 4),
-                  child: Text(
-                    '${entry.key}: ${entry.value}',
-                    style: TextStyle(
-                      fontSize: 14,
-                      color: IbTheme.textColor(context).withAlpha(204),
-                    ),
-                  ),
-                );
-              }).toList(),
-            ],
-          ),
-        );
-
-      case 'subsection':
-        final heading = item['heading'] as String? ?? '';
-        final tables = item['tables'] as List<dynamic>? ?? [];
-        return Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            if (heading.isNotEmpty)
-              Padding(
-                padding: const EdgeInsets.only(top: 12, bottom: 8),
-                child: Text(
-                  heading,
-                  style: TextStyle(
-                    fontSize: 17,
-                    fontWeight: FontWeight.w600,
-                    color: IbTheme.primaryHeadingColor(context),
-                  ),
-                ),
-              ),
-            ...tables.map((table) => _buildTable(context, table)).toList(),
-          ],
-        );
-
-      case 'steps':
-        final title = item['title'] as String? ?? 'Steps';
-        final steps = item['steps'] as List<dynamic>? ?? [];
-        final note = item['note'] as String?;
-        return Container(
-          margin: const EdgeInsets.only(bottom: 16),
-          padding: const EdgeInsets.all(16),
-          decoration: BoxDecoration(
-            color: IbTheme.textColor(context).withAlpha(13),
-            borderRadius: BorderRadius.circular(12),
-            border: Border.all(
-              color: IbTheme.textColor(context).withAlpha(26),
-            ),
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                title,
-                style: TextStyle(
-                  fontSize: 16,
-                  fontWeight: FontWeight.w600,
-                  color: IbTheme.primaryHeadingColor(context),
-                ),
-              ),
-              const SizedBox(height: 12),
-              ...steps.map((step) {
-                final stepNum = step['step'] as int? ?? 0;
-                final description = step['description'] as String? ?? '';
-                return Padding(
-                  padding: const EdgeInsets.only(bottom: 12),
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Container(
-                        width: 28,
-                        height: 28,
-                        decoration: BoxDecoration(
-                          color: IbTheme.getPrimaryColor(context),
-                          shape: BoxShape.circle,
-                        ),
-                        child: Center(
-                          child: Text(
-                            '$stepNum',
-                            style: const TextStyle(
-                              fontSize: 14,
-                              fontWeight: FontWeight.bold,
-                              color: Colors.white,
-                            ),
-                          ),
-                        ),
-                      ),
-                      const SizedBox(width: 12),
-                      Expanded(
-                        child: Text(
-                          description,
-                          style: TextStyle(
-                            fontSize: 14,
-                            height: 1.5,
-                            color: IbTheme.textColor(context).withAlpha(204),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                );
-              }).toList(),
-              if (note != null) ...[
-                const SizedBox(height: 8),
-                Container(
-                  padding: const EdgeInsets.all(12),
-                  decoration: BoxDecoration(
-                    color: IbTheme.getPrimaryColor(context).withAlpha(26),
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Icon(
-                        Icons.info_outline_rounded,
-                        size: 18,
-                        color: IbTheme.getPrimaryColor(context),
-                      ),
-                      const SizedBox(width: 8),
-                      Expanded(
-                        child: Text(
-                          note,
-                          style: TextStyle(
-                            fontSize: 13,
-                            color: IbTheme.textColor(context).withAlpha(204),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ],
-          ),
-        );
-
-      case 'widget':
-        final widgetType = item['widget_type'] as String?;
-        
-        if (widgetType == 'binary_simulator') {
-          return const Padding(
-            padding: EdgeInsets.symmetric(vertical: 16),
-            child: BinarySimulatorWidget(),
-          );
-        }
-        return const SizedBox.shrink();
-
-      case 'quiz':
-        final questionsRaw = item['questions'] as List<dynamic>? ?? [];
-        // Convert Map<dynamic, dynamic> to Map<String, dynamic>
-        final questions = questionsRaw.map((q) {
-          if (q is Map<String, dynamic>) {
-            return q;
-          } else {
-            return Map<String, dynamic>.from(q as Map);
+      // Load recommendations for topic page if not already loaded
+      if (!_loadingRecommendations && _recommendations.isEmpty) {
+        WidgetsBinding.instance.addPostFrameCallback((_) {
+          if (mounted && !_loadingRecommendations && _recommendations.isEmpty) {
+            print('NewIbContent: Calling _loadRecommendations from build for topic');
+            _loadRecommendations();
           }
-        }).toList();
-        return QuizWidget(questions: questions);
+        });
+      }
 
-      default:
-        return const SizedBox.shrink();
-    }
-  }
-
-  Widget _buildTable(BuildContext context, dynamic table) {
-    final title = table['title'] as String?;
-    final headers = table['headers'] as List<dynamic>?;
-    final rows = table['rows'] as List<dynamic>?;
-
-    if (headers == null || rows == null) {
-      return const SizedBox.shrink();
+      return IbTopicPage(
+        model: widget.model,
+        topicData: widget.model.topicData!,
+        recommendations: _recommendations,
+        loadingRecommendations: _loadingRecommendations,
+      );
     }
 
-    return Container(
-      margin: const EdgeInsets.symmetric(vertical: 12),
-      decoration: BoxDecoration(
-        border: Border.all(color: IbTheme.textColor(context).withAlpha(51)),
-        borderRadius: BorderRadius.circular(8),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          if (title != null)
-            Container(
-              width: double.infinity,
-              padding: const EdgeInsets.all(12),
-              decoration: BoxDecoration(
-                color: IbTheme.getPrimaryColor(context).withAlpha(26),
-                borderRadius: const BorderRadius.only(
-                  topLeft: Radius.circular(8),
-                  topRight: Radius.circular(8),
-                ),
-              ),
-              child: Text(
-                title,
-                style: Theme.of(context).textTheme.titleSmall?.copyWith(
-                  color: IbTheme.primaryHeadingColor(context),
-                  fontWeight: FontWeight.w600,
-                ),
-              ),
-            ),
-          Table(
-            border: TableBorder(
-              horizontalInside: BorderSide(
-                color: IbTheme.textColor(context).withAlpha(51),
-              ),
-              verticalInside: BorderSide(
-                color: IbTheme.textColor(context).withAlpha(51),
-              ),
-            ),
-            children: [
-              TableRow(
-                decoration: BoxDecoration(
-                  color: IbTheme.textColor(context).withAlpha(51),
-                ),
-                children:
-                    headers.map((header) {
-                      return Container(
-                        width: double.infinity,
-                        padding: const EdgeInsets.all(12),
-                        child: Text(
-                          header.toString(),
-                          style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                            color: IbTheme.primaryHeadingColor(context),
-                            fontWeight: FontWeight.w600,
-                          ),
-                        ),
-                      );
-                    }).toList(),
-              ),
-              ...rows.map((row) {
-                final rowData = row as List<dynamic>;
-                return TableRow(
-                  children:
-                      rowData.map((cell) {
-                        return Padding(
-                          padding: const EdgeInsets.all(12),
-                          child: Text(
-                            cell.toString(),
-                            style: Theme.of(context).textTheme.bodyMedium
-                                ?.copyWith(color: IbTheme.textColor(context)),
-                          ),
-                        );
-                      }).toList(),
-                );
-              }),
-            ],
-          ),
-        ],
-      ),
+    // Load recommendations for chapter index page if not already loaded
+    if (!_loadingRecommendations && _recommendations.isEmpty) {
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        if (mounted && !_loadingRecommendations && _recommendations.isEmpty) {
+          print('NewIbContent: Calling _loadRecommendations from build for chapter index');
+          _loadRecommendations();
+        }
+      });
+    }
+
+    // Show chapter index page
+    return IbChapterIndexPage(
+      model: widget.model,
+      chapter: widget.model.selectedChapter!,
+      recommendations: _recommendations,
+      loadingRecommendations: _loadingRecommendations,
     );
   }
-
-  Widget _buildCommentInput(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: IbTheme.textColor(context).withAlpha(13),
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(color: IbTheme.textColor(context).withAlpha(26)),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          TextField(
-            maxLines: 4,
-            decoration: InputDecoration(
-              hintText: 'Share your thoughts or ask a question...',
-              hintStyle: TextStyle(
-                color: IbTheme.textColor(context).withAlpha(128),
-              ),
-              border: InputBorder.none,
-            ),
-            style: TextStyle(color: IbTheme.textColor(context), fontSize: 16),
-          ),
-          const SizedBox(height: 12),
-          Align(
-            alignment: Alignment.centerRight,
-            child: ElevatedButton.icon(
-              onPressed: () {},
-              icon: const Icon(Icons.send_rounded, size: 18),
-              label: const Text('Post Comment'),
-              style: ElevatedButton.styleFrom(
-                backgroundColor: IbTheme.getPrimaryColor(context),
-                foregroundColor: Colors.white,
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 20,
-                  vertical: 12,
-                ),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildCommentsList(BuildContext context) {
-    final comments = [
-      {
-        'author': 'Santam',
-        'time': '2 hours ago',
-        'text': 'Great explanation! This really helped me understand binary representation.',
-        'likes': 12,
-      },
-      {
-        'author': 'Roy',
-        'time': '5 hours ago',
-        'text': 'Can someone explain the difference between signed and unsigned numbers?',
-        'likes': 5,
-      },
-      {
-        'author': 'Choudhury',
-        'time': '1 day ago',
-        'text': 'The examples are very clear. Thanks for this resource!',
-        'likes': 8,
-      },
-    ];
-
-    return Column(
-      children: comments
-          .map(
-            (comment) => _buildCommentCard(
-              context,
-              comment['author'] as String,
-              comment['time'] as String,
-              comment['text'] as String,
-              comment['likes'] as int,
-            ),
-          )
-          .toList(),
-    );
-  }
-
-  Widget _buildCommentCard(
-    BuildContext context,
-    String author,
-    String time,
-    String text,
-    int likes,
-  ) {
-    return Container(
-      margin: const EdgeInsets.only(bottom: 16),
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: IbTheme.textColor(context).withAlpha(13),
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(color: IbTheme.textColor(context).withAlpha(26)),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              CircleAvatar(
-                radius: 20,
-                backgroundColor: IbTheme.getPrimaryColor(context).withAlpha(51),
-                child: Text(
-                  author[0].toUpperCase(),
-                  style: TextStyle(
-                    color: IbTheme.getPrimaryColor(context),
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-              const SizedBox(width: 12),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      author,
-                      style: TextStyle(
-                        fontSize: 16,
-                        fontWeight: FontWeight.w600,
-                        color: IbTheme.primaryHeadingColor(context),
-                      ),
-                    ),
-                    Text(
-                      time,
-                      style: TextStyle(
-                        fontSize: 12,
-                        color: IbTheme.textColor(context).withAlpha(128),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 12),
-          Text(
-            text,
-            style: TextStyle(
-              fontSize: 15,
-              height: 1.5,
-              color: IbTheme.textColor(context).withAlpha(204),
-            ),
-          ),
-          const SizedBox(height: 12),
-          Row(
-            children: [
-              InkWell(
-                onTap: () {},
-                child: Row(
-                  children: [
-                    Icon(
-                      Icons.thumb_up_outlined,
-                      size: 18,
-                      color: IbTheme.textColor(context).withAlpha(179),
-                    ),
-                    const SizedBox(width: 6),
-                    Text(
-                      '$likes',
-                      style: TextStyle(
-                        fontSize: 14,
-                        color: IbTheme.textColor(context).withAlpha(179),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              const SizedBox(width: 24),
-              InkWell(
-                onTap: () {},
-                child: Row(
-                  children: [
-                    Icon(
-                      Icons.reply_rounded,
-                      size: 18,
-                      color: IbTheme.textColor(context).withAlpha(179),
-                    ),
-                    const SizedBox(width: 6),
-                    Text(
-                      'Reply',
-                      style: TextStyle(
-                        fontSize: 14,
-                        color: IbTheme.textColor(context).withAlpha(179),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
 }
