+++ lib/ui/views/new_ib/new_ib_landing_view.dart
@@ -3,7 +3,7 @@ import 'package:mobile_app/ib_theme.dart';
 import 'package:mobile_app/models/ib/new_ib_drawer_data.dart';
 import 'package:mobile_app/ui/views/base_view.dart';
 import 'package:mobile_app/ui/views/new_ib/components/new_ib_content.dart';
-import 'package:mobile_app/ui/views/new_ib/components/new_ib_drawer.dart';
+import 'package:mobile_app/ui/views/new_ib/shared/drawer/new_ib_drawer.dart';
 import 'package:mobile_app/viewmodels/ib/new_ib_landing_viewmodel.dart';
 
 class NewInteractiveBookView extends StatefulWidget {
