+++ lib/ui/views/new_ib/shared/widgets/binary_simulator_widget.dart
